requirements.txt
# packaging
pip>=9
setuptools>=26
twine
wheel>=0.29

# data science
numpy
pandas
scikit-learn
nltk
spacy
seaborn
keras==2.8.0
Keras-Preprocessing==1.1.2
tensorflow-text==2.8.1
tensorflow-estimator==2.8.0
tensorflow-model-optimization==0.7.1
tensorflow-hub==0.12.0
tf-models-official==2.8.0
tf-estimator-nightly==2.8.0.dev2021122109
python-multipart
gensim

# tests/linter
black
coverage
flake8
pytest
yapf

# API
gcsfs
google-cloud-storage
mlflow
s3fs
fastapi
uvicorn

# utilities
six>=1.14
joblib
memoized-property
termcolor
